Add floating point support into for..next loop

+++ src/math.c
@@ -1,9 +1,9 @@
-int max(int a, int b)
+double max(double a, double b)
 {
     return (a > b) ? a : b;
 }
 
-int min(int a, int b)
+double min(double a, double b)
 {
     return (a < b) ? a : b;
 }
